Convert a CLS end to end test to wrap a WPT.
This unifies MetricIntegrationTest.LayoutInstability with the nearly
identical simple-block-movement.html, and introduces an abstraction for
communicating expectations from the WPT to the wrapper, so that they can
be verified in a generic way against the trace data and UMA/UKM results.

We'll use this technique for CLSAttribution_* in a follow-up.

Bug: 1045064
Change-Id: I401af5cd0186ad3de930130694ff2fcf7b43ed7e