[FlexNG] Make ComputeMinMaxSizes cache more precise.
Previously our cache was imprecise, we recalculate our min/max sizes when
we didn't strictly need to, as we didn't know if we'd have a %-block-size
replaced element beneath us.

This patch changes ComputeMinMaxSizes to return a MinMaxSizesResult
(instead of just MinMaxSizes). This contains an additional bit of
information to indicate if this sub-tree depended on the %-block-size
given to it.

For most things, this will return false.

Along with this change, the cache changes in the following way:
 - We store an addition bit of information on LayoutBox to store the bit
   of information from ComputeMinMaxSizes.
 - We calculate the %-block-size for our children up-front, this means
   each of the algorithms don't have to do this.
 - With these changes, the cache can now hit, when the %-block-size given
   to us changes.

\o/

Bug: 845235
Change-Id: I93504721dbdd13bef58917b1fab2cabbcd1ce51e