Recalculate raster scales whenever the desired raster scale changes
Previous change for moving directly composited images from Blink to cc
missed the cases where the bounds changed in such a way that didn't trip
the code to recalculate raster scale. This means we were re-rastering
with an out of date scale, which can cause visible artifacts around
the edges of the images.

This change removes the misguided attempt to track raster scale aspect
ratio and instead uses the desired raster scale to determine when we
need to recalculate.

Bug:1075911

Change-Id: Iff0802429aee235ac4a557b181df9ec60a964ed6